+++ playwright-e2e/data-builders/exui/claimant-defendant-solicitor/evidence-upload-applicant/evidence-upload-applicant-data-builder-components.ts
@@ -1,8 +1,9 @@
-import DateHelper from '../../../../../helpers/date-helper';
-import { UploadDocumentValue } from '../../../../../models/ccd-case-data';
-import CaseDataHelper from '../../../../../helpers/case-data-helper';
-import { Party } from '../../../../../models/users/partys';
-import ClaimTrack from '../../../../../constants/cases/claim-track';
+import DateHelper from '../../../../helpers/date-helper';
+import { UploadDocumentValue } from '../../../../models/ccd-case-data';
+import CaseDataHelper from '../../../../helpers/case-data-helper';
+import ClaimTrack from '../../../../constants/cases/claim-track';
+import partys from '../../../../constants/users/partys';
+import ClaimType from '../../../../constants/cases/claim-type';
 
 const createDate = () =>
   DateHelper.formatDateToString(DateHelper.getToday(), { outputFormat: 'YYYY-MM-DD' });
@@ -12,20 +13,45 @@ const createDateTime = () =>
     outputFormat: 'YYYY-MM-DDTHH:MM:SS',
   });
 
-const evidenceUpload = {
-  EvidenceUpload: {
-    caseProgAllocatedTrack: 'FAST_CLAIM',
-  },
+const selectUploadOptions = (claimType: ClaimType) => {
+  if(claimType === ClaimType.TWO_VS_ONE) {
+    return {
+      SelectUploadOptions: {
+        evidenceUploadOptions: {
+          list_items: [CaseDataHelper.setCodeToData('Claimants 1 and 2')],
+          value: CaseDataHelper.setCodeToData('Claimants 1 and 2'),
+        },
+      }
+    }
+  }
+
+  return {};
+}
+
+const evidenceUpload = (claimTrack: ClaimTrack) => {
+  if(claimTrack === ClaimTrack.FAST_CLAIM)
+    return {
+      EvidenceUpload: {
+        caseProgAllocatedTrack: 'FAST_CLAIM',
+      },
+    };
+  else if (claimTrack === ClaimTrack.SMALL_CLAIM)
+    return {
+      EvidenceUpload: {
+        caseProgAllocatedTrack: 'SMALL_CLAIM',
+      }
+    };
+  return {};
 };
 
 const documentSelection = (claimTrack: ClaimTrack) => {
   if(claimTrack === ClaimTrack.FAST_CLAIM)
     return {
       DocumentSelectionFastTrack: {
-        disclosureSelectionEvidenceRes: ['DISCLOSURE_LIST'],
-        witnessSelectionEvidenceRes: ['WITNESS_SUMMARY'],
-        expertSelectionEvidenceRes: ['QUESTIONS_FOR_EXPERTS'],
-        trialSelectionEvidenceRes: ['COSTS'],
+        disclosureSelectionEvidence: ['DISCLOSURE_LIST'],
+        witnessSelectionEvidence: ['WITNESS_SUMMARY'],
+        expertSelectionEvidence: ['JOINT_STATEMENT'],
+        trialSelectionEvidence: ['DOCUMENTARY'],
       },
     };
   else if(claimTrack === ClaimTrack.SMALL_CLAIM)
@@ -35,78 +61,76 @@ const documentSelection = (claimTrack: ClaimTrack) => {
         trialSelectionEvidenceSmallClaim: ['AUTHORITIES'],
         expertSelectionEvidenceSmallClaim: ['EXPERT_REPORT'],
       }
-    };
-
+    }
   return {};
 };
 
+
 const documentUploadFastTrack = (
   claimTrack: ClaimTrack,
-  witnessParty: Party,
-  expertParty: Party,
   disclosureDocument: UploadDocumentValue,
   witnessSummaryDocument: UploadDocumentValue,
-  questionsDocument: UploadDocumentValue,
-  authoritiesDocument: UploadDocumentValue,
+  jointStatementDocument: UploadDocumentValue,
+  trialDocument: UploadDocumentValue
 ) => {
   if(claimTrack === ClaimTrack.FAST_CLAIM) {
-    const witnessPartyData = CaseDataHelper.buildWitnessData(witnessParty);
-    const expertPartyData = CaseDataHelper.buildExpertData(expertParty);
-
+    const witnessData = CaseDataHelper.buildWitnessData(partys.CLAIMANT_WITNESS_1);
+    const expertData = CaseDataHelper.buildExpertData(partys.CLAIMANT_EXPERT_1);
+  
     return {
       DocumentUpload: {
-        documentDisclosureListRes: [
+        documentDisclosureList: [
           CaseDataHelper.setIdToData({
             documentUpload: disclosureDocument,
             createdDatetime: createDateTime(),
           }),
         ],
-        documentWitnessSummaryRes: [
+        documentWitnessSummary: [
           CaseDataHelper.setIdToData({
-            witnessOptionName: witnessPartyData.partyName,
+            witnessOptionName: witnessData.partyName,
             witnessOptionUploadDate: createDate(),
             witnessOptionDocument: witnessSummaryDocument,
             createdDatetime: createDateTime(),
           }),
         ],
-        documentQuestionsRes: [
+        documentJointStatement: [
           CaseDataHelper.setIdToData({
-            expertOptionName: expertPartyData.partyName,
-            expertOptionOtherParty: 'text',
-            expertDocumentQuestion: 'question',
+            expertOptionName: expertData.partyName,
+            expertOptionExpertises: expertData.fieldOfExpertise,
             expertOptionUploadDate: createDate(),
-            expertDocument: questionsDocument,
+            expertDocument: jointStatementDocument,
             createdDatetime: createDateTime(),
           }),
         ],
-        documentAuthoritiesRes: [
+        documentEvidenceForTrial: [
           CaseDataHelper.setIdToData({
-            documentUpload: authoritiesDocument,
+            typeOfDocument: 'images etc',
+            documentIssuedDate: createDate(),
+            documentUpload: trialDocument,
             createdDatetime: createDateTime(),
           }),
         ],
       },
     };
   }
+
+  return {};
 };
 
 const documentUploadSmallClaim = (
   claimTrack: ClaimTrack,
-  witness1Party: Party,
-  witness2Party: Party,
-  expertParty: Party,
   witnessStatement1: UploadDocumentValue,
   witnessStatement2: UploadDocumentValue,
   expertReport: UploadDocumentValue,
   authoritiesDocument: UploadDocumentValue,
 ) => {
   if(claimTrack === ClaimTrack.SMALL_CLAIM) {
-    const witness1Data = CaseDataHelper.buildWitnessData(witness1Party);
-    const witness2Data = CaseDataHelper.buildWitnessData(witness2Party);
-    const expertData = CaseDataHelper.buildExpertData(expertParty);
+    const witness1Data = CaseDataHelper.buildWitnessData(partys.CLAIMANT_WITNESS_1);
+    const witness2Data = CaseDataHelper.buildWitnessData(partys.CLAIMANT_WITNESS_2);
+    const expertData = CaseDataHelper.buildExpertData(partys.CLAIMANT_EXPERT_1);
     return {
       DocumentUpload: {
-        documentWitnessStatementRes: [
+        documentWitnessStatement: [
           CaseDataHelper.setIdToData({
             witnessOptionName: witness1Data.partyName,
             witnessOptionUploadDate: createDate(),
@@ -118,7 +142,7 @@ const documentUploadSmallClaim = (
             witnessOptionDocument: witnessStatement2,
           }),
         ],
-        documentExpertReportRes: [
+        documentExpertReport: [
           CaseDataHelper.setIdToData({
             expertOptionName: expertData.partyName,
             expertOptionExpertise: expertData.fieldOfExpertise,
@@ -126,7 +150,7 @@ const documentUploadSmallClaim = (
             expertDocument: expertReport,
           }),
         ],
-        documentAuthoritiesRes: [
+        documentAuthorities: [
           CaseDataHelper.setIdToData({
             documentUpload: authoritiesDocument,
           }),
@@ -134,20 +158,12 @@ const documentUploadSmallClaim = (
       },
     };
   }
-}
-
-const undefine = {
-  Undefine: {
-    witnessSelectionEvidenceSmallClaim: undefined,
-    trialSelectionEvidenceSmallClaim: undefined,
-    expertSelectionEvidenceSmallClaim: undefined,
-  }
 };
 
 export default {
   evidenceUpload,
+  selectUploadOptions,
   documentSelection,
   documentUploadFastTrack,
   documentUploadSmallClaim,
-  undefine
 };
