--- playwright-e2e/data-builders/exui/claimant-defendant-solicitor/evidence-upload-respondent/unspec/evidence-upload-respondent-data-builder.ts
@@ -1,63 +1,0 @@
-import BaseDataBuilder from '../../../../../base/base-data-builder';
-import { defendantSolicitor1User } from '../../../../../config/users/exui-users';
-import ClaimTrack from '../../../../../constants/cases/claim-track';
-import partys from '../../../../../constants/users/partys';
-import { AllMethodsStep } from '../../../../../decorators/test-steps';
-import { Party } from '../../../../../models/users/partys';
-import evidenceUploadRespondentDataBuilderComponents from './evidence-upload-respondent-data-builder-components';
-
-@AllMethodsStep()
-export default class EvidenceUploadRespondentDataBuilder extends BaseDataBuilder {
-  async buildDS1Fast() {
-    return this.buildData({ claimTrack: ClaimTrack.FAST_CLAIM });
-  }
-
-  async buildDS1SmallClaim() {
-    return this.buildData();
-  }
-
-  protected async buildData(
-    {
-      claimTrack = ClaimTrack.SMALL_CLAIM,
-      witness1Party = partys.DEFENDANT_1_WITNESS_1,
-      witness2Party = partys.DEFENDANT_1_WITNESS_2,
-      expertParty = partys.DEFENDANT_1_EXPERT_1
-    } : {
-      claimTrack?: ClaimTrack,
-      witness1Party?: Party,
-      witness2Party?: Party,
-      expertParty?: Party
-    } = {}
-  ) {
-    const { civilServiceRequests } = this.requestsFactory;
-    const doc1 = await civilServiceRequests.uploadTestDocument(defendantSolicitor1User);
-    const doc2 = await civilServiceRequests.uploadTestDocument(defendantSolicitor1User);
-    const doc3 = await civilServiceRequests.uploadTestDocument(defendantSolicitor1User);
-    const doc4 = await civilServiceRequests.uploadTestDocument(defendantSolicitor1User);
-
-    return {
-      ...evidenceUploadRespondentDataBuilderComponents.evidenceUpload,
-      ...evidenceUploadRespondentDataBuilderComponents.documentSelection(claimTrack),
-      ...evidenceUploadRespondentDataBuilderComponents.documentUploadFastTrack(
-        claimTrack,
-        witness1Party,
-        expertParty,
-        doc1,
-        doc2,
-        doc3,
-        doc4,
-      ),
-      ...evidenceUploadRespondentDataBuilderComponents.documentUploadSmallClaim(
-        claimTrack,
-        witness1Party,
-        witness2Party,
-        expertParty,
-        doc1,
-        doc2,
-        doc3,
-        doc4,
-      ),
-      ...evidenceUploadRespondentDataBuilderComponents.undefine
-    };
-  }
-}
