+++ playwright-e2e/schema-builders/exui/claimant-defendant-solicitor/evidence-upload-applicant/unspec/evidence-upload-applicant-schema-builder.ts
@@ -12,6 +12,10 @@ export default class EvidenceUploadApplicantSchemaBuilder extends BaseSchemaBuil
     return this.buildSchema(caseDataBeforeSubmission, {claimTrack: ClaimTrack.FAST_CLAIM})
   }
 
+  async buildFast2v1(caseDataBeforeSubmission?: CCDCaseData) {
+    return this.buildSchema(caseDataBeforeSubmission, {claimTrack: ClaimTrack.FAST_CLAIM})
+  }
+
   async buildSmallClaim(caseDataBeforeSubmission?: CCDCaseData) {
     return this.buildSchema(caseDataBeforeSubmission);
   }
