+++ playwright-e2e/schema-builders/exui/claimant-defendant-solicitor/evidence-upload-respondent/unspec/evidence-upload-respondent-schema-builder.ts
@@ -12,6 +12,14 @@ export default class EvidenceUploadRespondentSchemaBuilder extends BaseSchemaBui
     return this.buildSchema(caseDataBeforeSubmission, {claimTrack: ClaimTrack.FAST_CLAIM})
   }
 
+  async buildDS1Fast1v2SS(caseDataBeforeSubmission?: CCDCaseData) {
+    return this.buildSchema(caseDataBeforeSubmission, {claimTrack: ClaimTrack.FAST_CLAIM})
+  }
+
+  async buildDS2Fast1v2SS(caseDataBeforeSubmission?: CCDCaseData) {
+    return this.buildSchema(caseDataBeforeSubmission, {claimTrack: ClaimTrack.FAST_CLAIM})
+  }
+
   async buildDS1SmallClaim(caseDataBeforeSubmission?: CCDCaseData) {
     return this.buildSchema(caseDataBeforeSubmission)
   }
